requirements.txt
discord.py
ezjsonpy
loguru
python-dotenv
pillow
requests